+++ connecteur-type/TdT/TdtRestamp.php
@@ -1,5 +1,7 @@
 <?php
 
+use Pastell\Step\AnnexeList;
+
 class TdTRestamp extends ConnecteurTypeActionExecutor
 {
     /**
@@ -12,7 +14,6 @@ class TdTRestamp extends ConnecteurTypeActionExecutor
         $tdt_error = $this->getMappingValue('tdt-error');
         $arrete_element = $this->getMappingValue('arrete');
         $acte_tamponne_element = $this->getMappingValue('acte_tamponne');
-        $autre_document_attache_element = $this->getMappingValue('autre_document_attache');
         $annexes_tamponnees_element = $this->getMappingValue('annexes_tamponnees');
         $acte_publication_date = $this->getMappingValue('acte_publication_date');
 
@@ -42,12 +43,29 @@ class TdTRestamp extends ConnecteurTypeActionExecutor
             $donneesFormulaire->addFileFromData($acte_tamponne_element, $actes_original_filename . '-tampon.pdf', $actes_tamponne);
         }
         if ($annexes_tamponnees_list) {
+            $annexes_envoyees = AnnexeList::getAll(
+                $this->getMappingValueList(AnnexeList::MAPPING_KEY),
+                $this->getDonneesFormulaire()
+            );
             $file_number = 0;
             foreach ($annexes_tamponnees_list as $i => $annexe_tamponnee) {
                 if (empty($annexe_tamponnee)) {
                     continue;
                 }
-                $annexe_filename_send = $tdT->getFilenameTransformation($this->getDonneesFormulaire()->getFileName($autre_document_attache_element, $i));
+                if (!isset($annexes_envoyees[$i])) {
+                    $message = 'Une erreur est survenue lors de la récupération des annexes tamponnées de ' . $tdT->getLogicielName() .
+                        " L'annexe tamponée " . $annexe_tamponnee['filename'] .
+                        " ne correspond à aucune des " . count($annexes_envoyees) . ' annexe(s) envoyée(s)';
+                    $this->changeOrUpdateAction($tdt_error, $message);
+                    $this->notify($tdt_error, $this->type, $message);
+                    return false;
+                }
+                $annexe_filename_send = $tdT->getFilenameTransformation(
+                    $this->getDonneesFormulaire()->getFileName(
+                        $annexes_envoyees[$i]['element'],
+                        $annexes_envoyees[$i]['num']
+                    )
+                );
                 if (strcmp($annexe_filename_send, $annexe_tamponnee['filename']) !== 0) {
                     $message = 'Une erreur est survenue lors de la récupération des annexes tamponnées de ' . $tdT->getLogicielName() .
                         " L'annexe tamponée " . $annexe_tamponnee['filename'] . ' ne correspond pas avec ' . $annexe_filename_send;
@@ -55,7 +73,10 @@ class TdTRestamp extends ConnecteurTypeActionExecutor
                     $this->notify($tdt_error, $this->type, $message);
                     return false;
                 }
-                $annexe_filename = $donneesFormulaire->getFileNameWithoutExtension($autre_document_attache_element, $i);
+                $annexe_filename = $donneesFormulaire->getFileNameWithoutExtension(
+                    $annexes_envoyees[$i]['element'],
+                    $annexes_envoyees[$i]['num']
+                );
                 $donneesFormulaire->addFileFromData(
                     $annexes_tamponnees_element,
                     $annexe_filename . '-tampon.pdf',